+++ src/cn/edu/zju/isst1/v2/restaurant/net/RestaurantResponse.java
@@ -25,14 +25,16 @@ public class RestaurantResponse extends CSTJsonResponse {
     @Override
     public void onResponse(JSONObject response) {
         super.onResponse(response);
-        if (clearDatabase) {
-            Lgr.i(response.toString());
-            CSTRestaurantDataDelegate.deleteAllRestaurent(mContext);
+        if (response != null) {
+            if (clearDatabase) {
+                Lgr.i(response.toString());
+                CSTRestaurantDataDelegate.deleteAllRestaurent(mContext);
 
-        }
-        CSTRestaurant restaurant = (CSTRestaurant) CSTJsonParser.parseJson(response, new CSTRestaurant());
-        for (CSTRestaurant restaurant_demo : restaurant.itemList) {
-            CSTRestaurantDataDelegate.saveRestaurant(mContext, restaurant_demo);
+            }
+            CSTRestaurant restaurant = (CSTRestaurant) CSTJsonParser.parseJson(response, new CSTRestaurant());
+            for (CSTRestaurant restaurant_demo : restaurant.itemList) {
+                CSTRestaurantDataDelegate.saveRestaurant(mContext, restaurant_demo);
+            }
         }
     }
 }
